+++ modules/devices.py
@@ -67,6 +67,12 @@ def has_triton(early:bool=False) -> bool:
     return test_triton(early=early)
 
 
+def get_hip_arch_name() -> str:
+    arch_name = getattr(torch.cuda.get_device_properties(device), "gcnArchName", "gfx0000")
+    arch_name = arch_name.split(':')[0]
+    return arch_name
+
+
 def get_backend(shared_cmd_opts):
     global args # pylint: disable=global-statement
     args = shared_cmd_opts
@@ -324,7 +330,7 @@ def test_fp16():
         elif backend == 'rocm':
             # gfx1102 (RX 7600, 7500, 7650 and 7700S) causes segfaults with fp16
             # agent can be overriden to gfx1100 to get gfx1102 working with ROCm so check the gpu name as well
-            agent = getattr(torch.cuda.get_device_properties(device), "gcnArchName", "gfx0000")
+            agent = get_hip_arch_name()
             agent_name = getattr(torch.cuda.get_device_properties(device), "name", "AMD Radeon RX 0000")
             if agent == "gfx1102" or (agent == "gfx1100" and any(i in agent_name for i in ("7600", "7500", "7650", "7700S"))):
                 fp16_ok = False
@@ -355,7 +361,7 @@ def test_bf16():
         elif backend == 'rocm' or backend == 'zluda':
             agent = None
             if backend == 'rocm':
-                agent = rocm.Agent(getattr(torch.cuda.get_device_properties(device), "gcnArchName", "gfx0000"))
+                agent = rocm.Agent(get_hip_arch_name())
             else:
                 from modules.zluda_installer import default_agent
                 agent = default_agent