+++ modules/rocm.py
@@ -83,9 +83,14 @@ class Agent:
             break
         return result
 
-    def __init__(self, name: str):
-        self.name = name
-        self.gfx_version = Agent.parse_gfx_version(name)
+    def __init__(self, name: str = None):
+        if name is None:
+            import torch
+            device = torch.cuda.current_device()
+            name = getattr(torch.cuda.get_device_properties(device), "gcnArchName", "gfx0000").split(':')[0]
+        else:
+            self.name = name.split(':')[0]
+        self.gfx_version = Agent.parse_gfx_version(self.name)
         if self.gfx_version > 0x1000:
             self.arch = MicroArchitecture.RDNA
         elif self.gfx_version in (0x908, 0x90a, 0x942,):
@@ -93,7 +98,7 @@ class Agent:
         else:
             self.arch = MicroArchitecture.GCN
         self.is_apu = (self.gfx_version & 0xFFF0 == 0x1150) or self.gfx_version in (0x801, 0x902, 0x90c, 0x1013, 0x1033, 0x1035, 0x1036, 0x1103,)
-        self.blaslt_supported = os.path.exists(os.path.join(blaslt_tensile_libpath, f"Kernels.so-000-{name}.hsaco" if sys.platform == "win32" else f"extop_{name}.co"))
+        self.blaslt_supported = os.path.exists(os.path.join(blaslt_tensile_libpath, f"Kernels.so-000-{self.name}.hsaco" if sys.platform == "win32" else f"extop_{name}.co"))
 
     @property
     def therock(self) -> Union[str, None]:
@@ -280,9 +285,9 @@ if sys.platform == "win32":
         try:
             import torch
             import numpy as np
-            from modules.devices import get_optimal_device
+            from modules.devices import get_hip_arch_name
 
-            gfx_version = Agent.parse_gfx_version(getattr(torch.cuda.get_device_properties(get_optimal_device()), "gcnArchName", "gfx0000"))
+            gfx_version = Agent.parse_gfx_version(get_hip_arch_name())
             if (gfx_version & 0xFFF0) == 0x1200:
                 # disable MIOpen for gfx120x
                 torch.backends.cudnn.enabled = False