+++ modules/shared_defaults.py
@@ -52,8 +52,7 @@ def get_default_modes(cmd_opts, mem_stat):
         default_sdp_override_choices.append('Triton Flash attention')
     elif devices.backend == "rocm":
         default_sdp_override_choices.append('Triton Flash attention')
-        import torch
-        if int(getattr(torch.cuda.get_device_properties(devices.device), "gcnArchName", "gfx0000")[3:]) < 1100:
+        if int(devices.get_hip_arch_name()) < 1100:
             default_sdp_override_options = ['Dynamic attention'] # only RDNA2 and older GPUs needs this
     elif devices.backend in {"directml", "cpu", "mps"}:
         default_sdp_override_options = ['Dynamic attention']